+++ eanalytics/src/main/java/com/eulerian/android/sdk/EAEstimate.java
@@ -56,7 +56,7 @@ public class EAEstimate extends EAProperties {
             return this;
         }
 
-        public Builder addProduct(Product product, int amount, int quantity) {
+        public Builder addProduct(Product product, double amount, int quantity) {
             if (quantity <= 0) {
                 EALog.w(EAEstimate.class.getSimpleName() + "#addProduct() : quantity might be > 0. Current is " +
                         quantity);
